refactor: assemble scheduled RTL in frontend
