+++ crates/celox-frontend-veryl/src/lib.rs
@@ -9,6 +9,7 @@ pub mod bitslicer;
 pub mod case;
 mod config;
 pub mod context_width;
+mod design_assembly;
 mod error;
 pub mod ff;
 pub mod flattening;
@@ -23,6 +24,7 @@ mod trace;
 mod types;
 
 pub use config::BuildConfig;
+pub use design_assembly::schedule_symbolic_rtl;
 pub use error::{LoweringPhase, ParserError, SourceLocation};
 pub use global_ff::{
     FfClockRecipe, FfRuntimeRelocation, SharedClockLowering, build_ff_clock_recipes,
