--- crates/celox/src/logic_tree.rs
@@ -1,5 +1,0 @@
-mod comb;
-pub use celox_slt::const_inline;
-pub use comb::SLTNode;
-pub use comb::{LogicPath, LogicPathTarget};
-pub use comb::{NodeId, SLTNodeArena, SLTNodeFactsError};
