--- crates/celox/src/logic_tree/comb.rs
@@ -1,1 +1,0 @@
-pub use celox_frontend_veryl::logic_tree::*;
