--- crates/celox/src/parser/scheduler.rs
@@ -1,1 +1,0 @@
-pub use celox_slt::scheduler::*;
